==> code/notebooks/1_combining_vgs.ipynb <==
import pandas as pd
import os

# Dictionary to convert Metacritic full platform names to VGChartz abbreviations
platform_dict = {
    "Game Boy Advance": "GBA", "DS": "DS", "PSP": "PSP", "Xbox One": "XOne", "Switch": "NS",
    "PlayStation 4": "PS4", "Wii U": "WiiU", "Xbox 360": "X360", "GameCube": "GC",
    "PlayStation": "PS", "Xbox Series X": "XS", "PlayStation 3": "PS3", "Dreamcast": "DC", "PlayStation 2": "PS2",
    "PC": "PC", "PlayStation 5": "PS5", "Wii": "Wii", "Xbox": "XB",
    "Nintendo 64": "N64", "PlayStation Vita": "PSV", "3DS": "3DS", "Stadia": "PC"
}

# Import all pages into a list
VGChartz = []
for i in range(6):
    curr_page = f"VGChartz_{i}.csv"
    inter = pd.read_csv(curr_page)
    VGChartz.append(inter)

# Create a sales data frame and concatenate all of the data frames in the list
sales = pd.concat(VGChartz, ignore_index=True)

# Create a new variable called games, which is the game name and aligns with Metacritic
sales['games'] = sales['game'].str.strip()

# Drop index and game column
drop = ["X", "game"]
sales = sales.drop(columns=[col for col in drop if col in sales.columns])

# Get the distinct values of sales
sales = sales.drop_duplicates()

# Import Metacritic data
Metacritic = pd.read_csv("Metacritic.csv")

# Make Metacritic's platform data look like VGChartz
Metacritic['platform'] = Metacritic['platform'].str.strip().map(platform_dict)

# Merge the sales and Metacritic data and only keep the values that appear in both
intersection = pd.merge(sales, Metacritic, on=["games", "platform"])

# Merge sales and Metacritic data with all of the data preserved
union = pd.merge(sales, Metacritic, on=["games", "platform"], how="outer")
union = union.sort_values(by=["user_score", "Total_Sales"], ascending=[False, False])

# Write them to their own separate files
union.to_csv("VGSales_Union.csv", index=False)
intersection.to_csv("VGSales_Intersect.csv", index=False)
